+++ configs/bootstrappers.php
@@ -3,6 +3,7 @@
  * Defines the list of bootstrapper classes to load
  */
 use Project\Bootstrappers\Databases\SQL;
+use Project\Bootstrappers\Events\Dispatcher;
 use Project\Bootstrappers\ORM\UnitOfWork;
 use RDev\Framework\Bootstrappers\Cryptography\Cryptography;
 
@@ -29,6 +30,7 @@ return [
      * List any console bootstrappers you'd like here
      * To enable Redis, add "Project\\Bootstrappers\\Databases\\Redis"
      */
+    Dispatcher::class,
     SQL::class,
     UnitOfWork::class
 ];