+++ configs/events.php
@@ -1,0 +1,21 @@
+<?php
+/**
+ * Defines the events config
+ */
+/**
+ * ----------------------------------------------------------
+ * List event names to their lists of listeners
+ * ----------------------------------------------------------
+ *
+ * Should be in the format:
+ *
+ * "eventName1" => [LIST_OF_LISTENERS],
+ * "eventName2" => [LIST_OF_LISTENERS]
+ *
+ * Listeners may either be callables or strings with the format:
+ *
+ * "Fully\\Qualified\\ClassName@methodName"
+ */
+return [
+
+];